requirements.txt
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
google-generativeai==0.3.2
pandas==2.2.0
requests==2.31.0
Pillow==10.2.0
linkedin-api==2.0.0a3 